format: - restructure file format - move all logic into query.ts
